Refactoring forecast generators and predictors for framework independence (#1052)
* fixed typing

* fixed typing

* added pytorch predictor class, unfinished

* added forecast gen

* changed dispatch function to dispatch on the prediction network type

* refactoring

* added example notebook and refactored imports

* fixed dependencies

* refactoring

* removed mxnet dependency from Predictor

* adapted tutorial

* fixed predictor usage

* edit predictor usage

* edit predictor usage

* merged shell/train.py

* added licence

* fix docstring

* fix docs

* cleared notebook cells

* removed pytorch predictor and example notebook

* added deprecation warning

* added Predictor and forecast_generator imports for backward compatibility

* removed unnecessary deprecation functions
