+++ docs/examples/extended_forecasting_tutorial/extended_tutorial.md
@@ -830,7 +830,8 @@ The estimator class is configured by a few hyperparameters and implements the re
 
 ```python
 from gluonts.model.estimator import GluonEstimator
-from gluonts.model.predictor import Predictor, RepresentableBlockPredictor
+from gluonts.model.predictor import Predictor
+from gluonts.mx.model.predictor import RepresentableBlockPredictor
 from gluonts.core.component import validated
 from gluonts.mx.trainer import Trainer
 from gluonts.support.util import copy_parameters
