+++ src/gluonts/model/deep_factor/_estimator.py
@@ -19,7 +19,8 @@ from gluonts import transform
 from gluonts.core.component import validated
 from gluonts.dataset.field_names import FieldName
 from gluonts.model.estimator import GluonEstimator
-from gluonts.model.predictor import Predictor, RepresentableBlockPredictor
+from gluonts.model.predictor import Predictor
+from gluonts.mx.model.predictor import RepresentableBlockPredictor
 from gluonts.mx.block.feature import FeatureEmbedder
 from gluonts.mx.distribution import DistributionOutput, StudentTOutput
 from gluonts.mx.trainer import Trainer
