+++ src/gluonts/model/deepstate/_estimator.py
@@ -25,7 +25,9 @@ from gluonts.core.component import validated
 from gluonts.dataset.field_names import FieldName
 from gluonts.model.deepstate.issm import ISSM, CompositeISSM
 from gluonts.model.estimator import GluonEstimator
-from gluonts.model.predictor import Predictor, RepresentableBlockPredictor
+from gluonts.model.predictor import Predictor
+from gluonts.mx.model.predictor import RepresentableBlockPredictor
+
 from gluonts.mx.distribution.lds import ParameterBounds
 from gluonts.mx.trainer import Trainer
 from gluonts.support.util import copy_parameters
