+++ src/gluonts/model/forecast_generator.py
@@ -12,31 +12,31 @@
 # permissions and limitations under the License.
 
 import logging
-from typing import Any, Callable, Iterator, List, Optional
+from typing import Callable, Iterator, List, Optional
+from functools import singledispatch
 
 # Third-party imports
-import mxnet as mx
 import numpy as np
 
+# First-party imports
 from gluonts.core.component import validated
 from gluonts.dataset.common import DataEntry
 from gluonts.dataset.field_names import FieldName
 from gluonts.dataset.loader import InferenceDataLoader
 from gluonts.model.forecast import (
-    DistributionForecast,
     Forecast,
     QuantileForecast,
     SampleForecast,
 )
 
-# First-party imports
-from gluonts.mx.distribution import DistributionOutput
-
 OutputTransform = Callable[[DataEntry, np.ndarray], np.ndarray]
-BlockType = mx.gluon.Block
-
 
 LOG_CACHE = set([])
+# different deep learning frameworks generate predictions and the tensor to numpy conversion differently,
+# use a dispatching function to prevent needing a ForecastGenerators for each framework
+@singledispatch
+def predict_to_numpy(prediction_net, tensor) -> np.ndarray:
+    raise NotImplementedError
 
 
 def log_once(msg):
@@ -44,34 +44,6 @@ def log_once(msg):
     if msg not in LOG_CACHE:
         logging.info(msg)
         LOG_CACHE.add(msg)
-
-
-def _extract_instances(x: Any) -> Any:
-    """
-    Helper function to extract individual instances from batched
-    mxnet results.
-
-    For a tensor `a`
-      _extract_instances(a) -> [a[0], a[1], ...]
-
-    For (nested) tuples of tensors `(a, (b, c))`
-      _extract_instances((a, (b, c)) -> [(a[0], (b[0], c[0])), (a[1], (b[1], c[1])), ...]
-    """
-    if isinstance(x, (np.ndarray, mx.nd.NDArray)):
-        for i in range(x.shape[0]):
-            # yield x[i: i + 1]
-            yield x[i]
-    elif isinstance(x, tuple):
-        for m in zip(*[_extract_instances(y) for y in x]):
-            yield tuple([r for r in m])
-    elif isinstance(x, list):
-        for m in zip(*[_extract_instances(y) for y in x]):
-            yield [r for r in m]
-    elif x is None:
-        while True:
-            yield None
-    else:
-        assert False
 
 
 class ForecastGenerator:
@@ -82,7 +54,7 @@ class ForecastGenerator:
     def __call__(
         self,
         inference_data_loader: InferenceDataLoader,
-        prediction_net: BlockType,
+        prediction_net,
         input_names: List[str],
         freq: str,
         output_transform: Optional[OutputTransform],
@@ -90,50 +62,6 @@ class ForecastGenerator:
         **kwargs
     ) -> Iterator[Forecast]:
         raise NotImplementedError()
-
-
-class DistributionForecastGenerator(ForecastGenerator):
-    @validated()
-    def __init__(self, distr_output: DistributionOutput) -> None:
-        self.distr_output = distr_output
-
-    def __call__(
-        self,
-        inference_data_loader: InferenceDataLoader,
-        prediction_net: BlockType,
-        input_names: List[str],
-        freq: str,
-        output_transform: Optional[OutputTransform],
-        num_samples: Optional[int],
-        **kwargs
-    ) -> Iterator[DistributionForecast]:
-        for batch in inference_data_loader:
-            inputs = [batch[k] for k in input_names]
-            outputs = prediction_net(*inputs)
-            if output_transform is not None:
-                outputs = output_transform(batch, outputs)
-            if num_samples:
-                log_once(
-                    "Forecast is not sample based. Ignoring parameter `num_samples` from predict method."
-                )
-
-            distributions = [
-                self.distr_output.distribution(*u)
-                for u in _extract_instances(outputs)
-            ]
-
-            i = -1
-            for i, distr in enumerate(distributions):
-                yield DistributionForecast(
-                    distr,
-                    start_date=batch["forecast_start"][i],
-                    freq=freq,
-                    item_id=batch[FieldName.ITEM_ID][i]
-                    if FieldName.ITEM_ID in batch
-                    else None,
-                    info=batch["info"][i] if "info" in batch else None,
-                )
-            assert i + 1 == len(batch["forecast_start"])
 
 
 class QuantileForecastGenerator(ForecastGenerator):
@@ -144,7 +72,7 @@ class QuantileForecastGenerator(ForecastGenerator):
     def __call__(
         self,
         inference_data_loader: InferenceDataLoader,
-        prediction_net: BlockType,
+        prediction_net,
         input_names: List[str],
         freq: str,
         output_transform: Optional[OutputTransform],
@@ -153,7 +81,7 @@ class QuantileForecastGenerator(ForecastGenerator):
     ) -> Iterator[Forecast]:
         for batch in inference_data_loader:
             inputs = [batch[k] for k in input_names]
-            outputs = prediction_net(*inputs).asnumpy()
+            outputs = predict_to_numpy(prediction_net, inputs)
             if output_transform is not None:
                 outputs = output_transform(batch, outputs)
 
@@ -185,7 +113,7 @@ class SampleForecastGenerator(ForecastGenerator):
     def __call__(
         self,
         inference_data_loader: InferenceDataLoader,
-        prediction_net: BlockType,
+        prediction_net,
         input_names: List[str],
         freq: str,
         output_transform: Optional[OutputTransform],
@@ -194,14 +122,14 @@ class SampleForecastGenerator(ForecastGenerator):
     ) -> Iterator[Forecast]:
         for batch in inference_data_loader:
             inputs = [batch[k] for k in input_names]
-            outputs = prediction_net(*inputs).asnumpy()
+            outputs = predict_to_numpy(prediction_net, inputs)
             if output_transform is not None:
                 outputs = output_transform(batch, outputs)
             if num_samples:
                 num_collected_samples = outputs[0].shape[0]
                 collected_samples = [outputs]
                 while num_collected_samples < num_samples:
-                    outputs = prediction_net(*inputs).asnumpy()
+                    outputs = predict_to_numpy(prediction_net, inputs)
                     if output_transform is not None:
                         outputs = output_transform(batch, outputs)
                     collected_samples.append(outputs)
@@ -223,3 +151,8 @@ class SampleForecastGenerator(ForecastGenerator):
                     info=batch["info"][i] if "info" in batch else None,
                 )
             assert i + 1 == len(batch["forecast_start"])
+
+
+# import mxnet-dependent DistributionForecastGenerator for backwards compatibility
+# TODO add deprecation warning
+from gluonts.mx.model.forecast_generator import DistributionForecastGenerator
