+++ src/gluonts/model/gp_forecaster/_estimator.py
@@ -22,7 +22,8 @@ from mxnet.gluon import HybridBlock
 from gluonts.core.component import DType, validated
 from gluonts.dataset.field_names import FieldName
 from gluonts.model.estimator import GluonEstimator
-from gluonts.model.predictor import Predictor, RepresentableBlockPredictor
+from gluonts.model.predictor import Predictor
+from gluonts.mx.model.predictor import RepresentableBlockPredictor
 from gluonts.mx.kernels import KernelOutput, RBFKernelOutput
 from gluonts.mx.trainer import Trainer
 from gluonts.support.util import copy_parameters
