+++ src/gluonts/model/simple_feedforward/_estimator.py
@@ -21,7 +21,7 @@ from mxnet.gluon import HybridBlock
 from gluonts.core.component import validated
 from gluonts.dataset.field_names import FieldName
 from gluonts.model.estimator import GluonEstimator
-from gluonts.model.predictor import RepresentableBlockPredictor
+from gluonts.mx.model.predictor import RepresentableBlockPredictor
 from gluonts.mx.distribution import DistributionOutput, StudentTOutput
 from gluonts.mx.trainer import Trainer
 from gluonts.transform import (
@@ -31,7 +31,7 @@ from gluonts.transform import (
     Transformation,
     AddObservedValuesIndicator,
 )
-from gluonts.model.forecast_generator import DistributionForecastGenerator
+from gluonts.mx.model.forecast_generator import DistributionForecastGenerator
 from gluonts.transform.feature import (
     DummyValueImputation,
     MissingValueImputation,
