+++ test/paper_examples/test_axiv_paper_examples.py
@@ -58,7 +58,8 @@ def test_appendix_c():
     from typing import List
     from mxnet import gluon
     from gluonts.model.estimator import GluonEstimator
-    from gluonts.model.predictor import Predictor, RepresentableBlockPredictor
+    from gluonts.model.predictor import Predictor
+    from gluonts.mx.model.predictor import RepresentableBlockPredictor
     from gluonts.mx.trainer import Trainer
     from gluonts.transform import (
         InstanceSplitter,
